id찾기 temptoken 유효문제

+++ lyc_flutter_project/lib/Join/Provider/login_provider.dart
@@ -4,16 +4,27 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:lyc_flutter_project/Join/model/Credential.dart';
 import 'package:lyc_flutter_project/common/const/data.dart';
 import 'package:lyc_flutter_project/common/dio/dio.dart';
-import '../Screens/login_screen.dart';
+import 'package:lyc_flutter_project/mypage/model/profile.dart';
+import 'package:lyc_flutter_project/mypage/repository/mypage_repository.dart';
 
 class LoginProvider extends ChangeNotifier {
   final DioProvider dioProvider;
   late final Dio dio;
   late final FlutterSecureStorage storage;
+  final MypageRepository mypageRepository;
 
   bool _isLoading = false;
   bool _isLoggedIn = false;
   int? _memberId;
+
+  Profile? _profile;
+  bool _hasProfile = false;
+
+  static const String memberIdKey = 'jaskfjajesfn';
+
+  get profile => _profile;
+
+  get hasProfile => _hasProfile;
 
   bool get isLoading => _isLoading;
 
@@ -21,16 +32,51 @@ class LoginProvider extends ChangeNotifier {
 
   int? get memberId => _memberId;
 
-  LoginProvider(this.dioProvider){
+  LoginProvider(this.dioProvider,
+      this.mypageRepository,) {
     dio = dioProvider.dio;
     storage = dioProvider.storage;
+    _loadMemberId();
+  }
+
+  Future<void> _loadMemberId() async {
+    final storedMemberId = await storage.read(key: memberIdKey);
+    if (storedMemberId != null) {
+      _memberId = int.parse(storedMemberId);
+      notifyListeners();
+    }
+  }
+
+  Future<void> _saveMemberId(int memberId) async {
+    await storage.write(key: memberIdKey, value: memberId.toString());
+    _memberId = memberId;
+    notifyListeners();
+  }
+
+  Future<bool> getProfile() async {
+    if (_memberId == null) return false;
+
+    try {
+      final resp = await mypageRepository.getProfile(memberId: _memberId!);
+
+      if (resp.isSuccess) {
+        _profile = resp.result;
+        _hasProfile = true;
+        notifyListeners();
+        return true;
+      } else {
+        return false;
+      }
+    } catch (e) {
+      print("getProfile 에러 발생: $e");
+      return false;
+    }
   }
 
   Future<void> login(String id, String pw, BuildContext context) async {
     _setLoading(true);
     try {
       final requestBody = Credential(loginId: id, loginPw: pw);
-
       final response = await dio.post(
         'http://$ip/lyc/auths/sign-in',
         data: requestBody,
@@ -51,9 +97,9 @@ class LoginProvider extends ChangeNotifier {
 
         await storage.write(key: refreshTokenKey, value: refreshToken);
         await storage.write(key: accessTokenKey, value: accessToken);
+        await _saveMemberId(memberId);
 
         _isLoggedIn = true;
-        _memberId = memberId;
         notifyListeners();
       } else {
         _showErrorDialog(context, '로그인 실패', 'API 요청이 실패했습니다.');
@@ -99,19 +145,22 @@ class LoginProvider extends ChangeNotifier {
     );
   }
 
-  Future<void> logout(BuildContext context) async {
+  Future<void> logout() async {
     try {
-      await storage.delete(key: refreshTokenKey);
-      await storage.delete(key: accessTokenKey);
-      _isLoggedIn = false;
-      _memberId = null;
-      notifyListeners();
-
-      Navigator.of(context).pushReplacement(
-        MaterialPageRoute(
-          builder: (context) => LoginScreen(),
-        ),
+      final resp = await dio.post(
+        'http://$ip/lyc/auths/sign-out',
+        options: Options(headers: {"accessToken": "true"}),
       );
+      if (resp.statusCode == 200) {
+        await storage.delete(key: refreshTokenKey);
+        await storage.delete(key: accessTokenKey);
+        await storage.delete(key: memberIdKey);
+        _isLoggedIn = false;
+        _memberId = null;
+        _profile = null;
+        _hasProfile = false;
+        notifyListeners();
+      }
     } catch (e) {
       print("Error in logout: $e");
     }
@@ -120,12 +169,13 @@ class LoginProvider extends ChangeNotifier {
   Future<void> checkLoginStatus() async {
     final refreshToken = await storage.read(key: refreshTokenKey);
     final accessToken = await storage.read(key: accessTokenKey);
-
-    if (refreshToken != null && accessToken != null) {
+    final storedMemberId = await storage.read(key: memberIdKey);
+    if (refreshToken != null && accessToken != null && storedMemberId != null) {
       _isLoggedIn = true;
+      _memberId = int.parse(storedMemberId);
     } else {
-      _isLoggedIn = false;
       _memberId = null;
+      _isLoggedIn = false;
     }
     notifyListeners();
   }
